File Upload and Download

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator')
 const Product = require('../models/product')
+const deleteFile = require('../util/file')
 
 exports.getAdminProductsPage = (req, res, next) => {
   Product.find({ userId: req.user._id })
@@ -62,9 +63,23 @@ exports.getEditProductPage = (req, res, next) => {
 
 exports.postProduct = (req, res, next) => {
   const title = req.body.title
-  const imageUrl = req.body.imageUrl
+  // const imageUrl = req.body.imageUrl
+  const image = req.file
+  console.log(image)
   const price = req.body.price
   const description = req.body.description
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      product: { title, price, description },
+      hasError: true,
+      errorMessage: 'Attached file is not an image!',
+      validationErrors: [],
+    })
+  }
 
   const errors = validationResult(req)
 
@@ -73,12 +88,14 @@ exports.postProduct = (req, res, next) => {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
-      product: { title, imageUrl, price, description },
+      product: { title, price, description },
       hasError: true,
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     })
   }
+
+  const imageUrl = image.path
 
   const product = new Product({
     title,
@@ -94,7 +111,7 @@ exports.postProduct = (req, res, next) => {
     })
     .catch((err) => {
       // return res.redirect('/500')
-      const error = new Error('Creating a product faild.')
+      const error = new Error(err)
       error.httpStatusCode = 500
       return next(error)
     })
@@ -103,7 +120,8 @@ exports.postProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.id
   const updatedTitle = req.body.title
-  const updatedImageUrl = req.body.imageUrl
+  // const updatedImageUrl = req.body.imageUrl
+  const updatedImage = req.file
   const updatedPrice = req.body.price
   const updatedDescription = req.body.description
 
@@ -116,7 +134,6 @@ exports.postEditProduct = (req, res, next) => {
       editing: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId,
@@ -139,7 +156,10 @@ exports.postEditProduct = (req, res, next) => {
         return res.redirect('/')
       }
       product.title = updatedTitle
-      product.imageUrl = updatedImageUrl
+      if (updatedImage) {
+        deleteFile(product.imageUrl)
+        product.imageUrl = updatedImage.path
+      }
       product.price = updatedPrice
       product.description = updatedDescription
       return product.save().then((result) => {
@@ -156,8 +176,15 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.id
 
-  // Product.findByIdAndDelete(prodId)
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('No product found.'))
+      }
+      deleteFile(product.imageUrl)
+      // Product.findByIdAndDelete(prodId)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then((result) => {
       console.log('DELETED PRODUCT!!!')
       res.redirect('/admin/products')
